Simple file storage Started work on legacy connectors

+++ rooibos/solr/models.py
@@ -13,7 +13,18 @@ class RecordInfo(models.Model):
     last_index = models.DateTimeField(null=True)
     
 
+DISABLE_SOLR_UPDATES = False
+
+def DisableSolrUpdates():
+    DISABLE_SOLR_UPDATES = True
+    
+def EnableSolrUpdates():
+    DISABLE_SOLR_UPDATES = False
+
+
 def post_delete_callback(sender, **kwargs):
+    if DISABLE_SOLR_UPDATES:
+        return
     try:
         id = kwargs['instance'].id
         RecordInfo.objects.filter(record__id=id).delete()
@@ -23,6 +34,8 @@ def post_delete_callback(sender, **kwargs):
         pass
     
 def post_save_callback(sender, **kwargs):
+    if DISABLE_SOLR_UPDATES:
+        return
     try:
         id = kwargs['instance'].id
         RecordInfo.objects.filter(record__id=id).delete()
@@ -30,8 +43,8 @@ def post_save_callback(sender, **kwargs):
         pass
     
 
-post_delete.connect(post_delete_callback, sender=Record)
-post_save.connect(post_save_callback, sender=Record)
+#post_delete.connect(post_delete_callback, sender=Record)
+#post_save.connect(post_save_callback, sender=Record)
 
 
 class SolrIndex():
